Move graphics::Image to its own file
